Create checkPintos
This file can be used with the Pintos Operating System to run tests while you are still coding and checking things. To do this add this file to your ~/bin directory and add the bin directory to your .bash_profile $PATH variable. Make sure to run bash before running checkPintos and also make sure to add the executable flag.